Refactor location of argument parsing logic
